3 template code files for identifying signs
I'm very sorry with how late this is. It doesn't work but has the framework to work when the bugs are fixed and patched out. I've been dealing with illness and many personal issues as of late that's been weighing me down that I would prefer not to get into.

As for the files uploaded:
3 template files I was trying to get working so we could use our own dataset and train with our own images for the UK sign layout. The dataset for the template is amongst these uploaded as 'dataset.zip'.  I would not recommend submitting this at all as of yet. There are so many bugs which I've documented and that's probably not all of them. They can probably be solved if the right packages are added. 
Index:
'Car Identifying Challenge.py' this was my first run through it so I kept saving as I was going on and needs much debugging, it's the current file I'm using.
'Car Identifying Challenge (GUI).py' is the GUI file for the project so we can test the AI
'Car Identifying Challenge (Entirety).py' is the current file I'm using as mentioned above all in one block of huge code
'Bugs.txt' discovered bugs throughout project in a text file. 
I'll see if I can place the dataset of images in a Discord chat
And yes, this is the template from the notes section in the discord